feat: add condition for work changr language
