+++ src/virtual.js
@@ -1,4 +1,4 @@
-import {keyValue, keyValueCaps,keyValueShift,/* keyValueRus,*/keyValueCapsShift } from './data.js'
+
 //---------------------------------work capsLock
 document.querySelector('#CapsLock').addEventListener('click', (e)=>{
 	console.log(e.target)
@@ -16,33 +16,41 @@ document.querySelector('#CapsLock').addEventListener('click', (e)=>{
 })
 document.querySelectorAll('.key-button')
 .forEach(el=>el.addEventListener('mousedown', (e)=>{
+
 		if(e.target.id !== 'CapsLock' ||  !e.target.classList.contains('capslock')){
 			document.querySelector(`#${e.target.id}`).classList.add('press');
 		}
+
 		if(e.target.id === "ShiftLeft" ||e.target.id === "ShiftLeft"){
 			document.querySelectorAll('.main').forEach(el=>{
 				el.classList.add('hidden')
-			})
+			});
+
 			document.querySelectorAll('.shift').forEach(el=>{
 				el.classList.remove('hidden')
-			})
-		
+			});
 		}
+
 		let capsLock = document.querySelector('.capslock')
 		if((e.target.id === "ShiftLeft" ||e.target.id === "ShiftLeft") && capsLock.classList.contains('press')){
+			
 			document.querySelectorAll('.shift').forEach(el=>{
-				el.classList.add('hidden')
+				el.classList.add('hidden');
 			});
+
 			document.querySelectorAll('.caps').forEach(el=>{
-				el.classList.toggle('hidden')
+				el.classList.toggle('hidden');
 			});
+
 			document.querySelectorAll('.capsShift').forEach(el=>{
-				el.classList.remove('hidden')
+				el.classList.remove('hidden');
 			})
 		}
 }));
+
 document.querySelectorAll('.key-button')
 .forEach(el=>el.addEventListener('mouseup', (e)=>{
+
 	if(e.target.id !== 'CapsLock' ||  !e.target.classList.contains('capslock')){
 		document.querySelector(`#${e.target.id}`).classList.remove('press');
 	}
@@ -50,26 +58,28 @@ document.querySelectorAll('.key-button')
 	if(e.target.id === "ShiftLeft" || e.target.id === "ShiftLeft"){
 		document.querySelectorAll('.main').forEach(el=>{
 			el.classList.remove('hidden')
-		})
+		});
+
 		document.querySelectorAll('.shift').forEach(el=>{
 			el.classList.add('hidden')
 		})
-		// document.querySelector('#CapsLock').classList.toggle('press')
 	}
+
 	let capsLock = document.querySelector('.capslock')
 		if((e.target.id === "ShiftLeft" ||e.target.id === "ShiftLeft") && capsLock.classList.contains('press')){
+			
 			document.querySelectorAll('.main').forEach(el=>{
 				el.classList.add('hidden')
 			});
+
 			document.querySelectorAll('.caps').forEach(el=>{
 				el.classList.toggle('hidden')
 			});
+
 			document.querySelectorAll('.capsShift').forEach(el=>{
 				el.classList.add('hidden')
 			})
-		}
-	
-	
+		}	
 }));
